added horizontal motion based on phone tilt
sensors dependency was included for accelerometer readings.

subscription line was added to avoid printing infinite accelerometer print statements in console. It was a pain in the ass.

I have set the gravity, vY and vX values that seem to work well. These may need to be changed when obstacles are added.

renamed variables to lower camel case.

accelerometer button and page were added for testing. will be removed later on.